Added API versioning - Migrated current endpoints to routes/v1 directory
